refactor(cli): split clipboard backends
